feat: users can no longer update or delete others habits, checking the user field on the habit matches the user.id from the auth headers

+++ backend/controllers/habits.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Habits = require("../models/habits.js");
+const User = require("../models/user.js");
 const mongoose = require("mongoose");
 
 const getAllHabits = asyncHandler(async (req, res) => {
@@ -35,6 +36,18 @@ const updateHabit = asyncHandler(async (req, res) => {
   }
 
   const habit = req.body;
+  const habitCheck = await Habits.findById(req.params.id);
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (habitCheck.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
 
   const updatedHabit = await Habits.findByIdAndUpdate(
     _id,
@@ -49,7 +62,18 @@ const deleteHabit = asyncHandler(async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
     throw new Error(`No habit with ID ${req.params.id} found`);
   }
+  const habit = await Habits.findById(req.params.id);
 
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (habit.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   await Habits.findByIdAndRemove(req.params.id);
   res.json({ message: `Succesfully deleted Habit with ID ${req.params.id}` });
 });
